requirements.txt
cffi==1.16.0
cryptography==42.0.7
pillow==10.3.0
pycparser==2.22
sqlcipher3-wheels==0.5.2.post1
